requirements.txt
discord.py==2.5.2
aiohttp==3.12.13
python-dotenv==1.0.1